requirements.txt
customtkinter
mysql-connector-python
pandas
scikit-learn
numpy